feat: update device flow to draft-12
User Code inputs can now be mixed case and punctuation and whitespace is
ignored. To  accommodate for this quality of life feature the `charset`
is now fixed to be either base-20 or just digits and `mask` can only
include `*- ` characters. DeviceCode model's userCode property is now
always uppercase with no punctuation or whitespace.

The rendered text input is now opened with a numeric keyboard on mobile
devices if the charset is `digits`.