feat: check session client Origin check
To satisfy a normative requirement the Origin of a valid formatted
postMessage must be checked to be a valid/expected one.

The iframe will now do this check with an XMLHttpRequest to the provider
with the client_id and Origin and will check the Origin to be amongst
the Origins of a given client's registered redirect_uris. This is enough
since the session state computation uses the redirect_uri Origin.

If the client is not found, request is malformed or Origin is not
whitelisted the frame will respond with "error" status back.

Additionally: more format checks are in place to filter out invalid
messages from i.e. trackers.

Additionally: the check_session_iframe debug was removed as it was not
sufficient for debugging anyway.