feat: allow for client default metadata to be changed
To support OIDC profiles such as Financial-grade API which e.g. use
PS256 as default algorithm and forbid the use of RS256 this
configuration option enables deployments that do not conform to the
existing OIDC profiles but are intended to be used for others.

Also to e.g. change the default client authentication method